Pending changes exported from your codespace

+++ src/pbhelper/pbhelper_DrawRectangleLinesEx.c
@@ -1,5 +1,5 @@
 ﻿#include "raylib_pb_helper.h"
 
-void pbhelper_DrawRectangleLinesEx(Rectangle* rect, int lineThick, Color color) {
+void pbhelper_DrawRectangleLinesEx(Rectangle* rect, float lineThick, Color color) {
     if( rect ) DrawRectangleLinesEx(*rect, lineThick, color);
 }
